requirements.txt
jinja2
requests
jsonschema
nose
mock
pyopenssl
DeepDiff
mock
xmljson
JSONBender
pyyaml